cluster: ready the localization white-box layer-views run for the full 13,769-answer population
Omri asked for white-box data on everything we currently evaluate — ProcessBench
(Qwen3-4B + Qwen3-8B) and PRMBench (Qwen3-8B), 13,769 answers / 6,968,779 tokens —
so the per-layer lens field can be fused with the token-level entropy bank. Attention
and hidden states were never captured for this population (the one cell whose name
suggests otherwise has an empty `capture` manifest), so this is a new extraction.

`run_localization_layer_views.py` and its sbatch wrapper were written for exactly this
population but had never run. Four fixes before a 7M-token job:

* cov_eigs stored as float32, not float16. Residual Gram eigenvalues routinely exceed
  65504 on a trained model; DATA_INVENTORY.md records 47,008 non-finite entries on the
  Qwen3-8B cell from this overflow.
* cov_eigs and hid_proj are now kept. candidate_layer_field computed them on every row
  and the driver discarded them, paying the eigendecomposition for nothing and putting
  the INSIDE/EigenScore/HaloScope geometry family out of reach of this output.
* A vacuous gate now aborts. ProcessBench rows key on "id" and PRMBench rows on "idx";
  a join miss would otherwise pass as success with zero rows actually checked.
* `--out` is validated before it is dereferenced, and the manifest carries proj_dim,
  cov_eigs_r, batch caps, layer count, hidden size, job id and git sha.

Defaults raised to the full field (attn,mlp,resid; r=32; D=256): ~944 B/token, ~6.9 GB
for the whole population. Disk is not the constraint — the lens runs norm+lm_head 108
times per token, so GPU time is.

Also documents the AIRCC Slurm trap: `ssh aircc '<cmd>'` is a non-login shell and never
sources /etc/profile.d, so SLURM_CONF_SERVER is unset and every squeue/sbatch dies with
"Could not establish a configuration source". It reads as an outage and is not one.

Local CPU smoke passes at the new defaults.

+++ cluster/layer_lens.py
@@ -190,7 +190,11 @@ def candidate_layer_field(mdl, tap, hidden_states, gen_ids, plen, tgen, hid_proj
 
     field = {q: np.empty((len(MODULES), L, tgen), dtype=np.float16) for q in QUANTITIES}
     resid_norm = np.empty((L, tgen), dtype=np.float16)
-    cov_eigs = np.zeros((L, cov_eigs_r), dtype=np.float16)
+    # float32, NOT float16: residual Gram eigenvalues routinely exceed the float16 max
+    # (65504) on trained models — results/whitebox_layer_fusion_v2/DATA_INVENTORY.md
+    # records 47,008 non-finite cov_eigs entries on the Qwen3-8B cell from exactly this
+    # overflow.  The array is [L, r] per candidate, so float32 costs nothing.
+    cov_eigs = np.zeros((L, cov_eigs_r), dtype=np.float32)
     proj = np.empty((L, hid_proj.shape[1]), dtype=np.float16)
 
     # ``hidden_states[L]`` IS NOT x_L.  HF applies the final norm before appending the
@@ -235,7 +239,7 @@ def candidate_layer_field(mdl, tap, hidden_states, gen_ids, plen, tgen, hid_proj
             # far smaller than d here, so the T x T form is the cheap one.
             gram = (xc @ xc.T) / max(tgen - 1, 1)
             ev = torch.linalg.eigvalsh(gram.double()).flip(0)[:cov_eigs_r]
-            cov_eigs[l, :ev.numel()] = ev.cpu().numpy().astype(np.float16)
+            cov_eigs[l, :ev.numel()] = ev.cpu().numpy().astype(np.float32)
 
     del ref_lp
     out = {q: field[q] for q in QUANTITIES}
